MindMeld--AI-CBT(mod 2)/MindMeld--AI-CBT(mod 2)/requirement.txt
flask_cors
google.generativeai
dotenv